refactor: 서버에서 html 파일 만드는 부분 이동
