--- src/client/App.server.tsx
@@ -1,17 +1,0 @@
-import ReactDOMServer from "react-dom/server";
-import { ServerStyleSheet } from "styled-components";
-
-import App from "./App";
-
-const styleSheet = new ServerStyleSheet();
-const render = (): { html: string; styleTags: string } => {
-  const html: string = ReactDOMServer.renderToString(
-    styleSheet.collectStyles(<App />)
-  );
-
-  const styleTags = styleSheet.getStyleTags();
-
-  return { html, styleTags };
-};
-
-export default render;
